Fix MISRAC for the folder pm/drivers/clock
 - MISRA.ETYPE.ASSIGN.2012
  - The value of an expression shall not be assigned
    to an object with a narrower essential type or of
    a different essential type category
  - Fixed by having the required datatype

 - MISRA.ETYPE.CATEGORY.DIFFERENT.2012
  - Both operands of an operator in which the usual
    arithmetic conversions are performed shall have
    the same essential type category
  - Fixed by typecasting to the required datatype

 - MISRA.ETYPE.COMP.CAST.IMPL.WIDER.2012
  - If a composite expression is used as one operand of
    an operator in which the usual arithmetic conversions
    are performed then the other operand shall not have wider
    essential type
  - Fixed by typecasting to the required datatype

 - MISRA.ETYPE.INAPPR.OPERAND.UNOP.2012
  - Operands shall not be of an inappropriate essentialtype
  - Fixed by having a boolean expression in the condition

 - MISRA.EXPR.PARENS.2012
  - The precedence of operators within expressions should
    be made explicit
  - Fixed by adding the required braces

 - MISRA.GOTO
  - The goto statement shall not be used.
  - Fixed by a new variable for return values/expressions

 - MISRA.RETURN.NOT_LAST
  - A function should have a single point of exit at the end
  - Fixed by a new variable for return values/expressions

 - MISRA.STMT.COND.NOT_BOOLEAN.2012
  - The controlling expression of an if statement and the
    controlling expression of an iteration-statement shall
    have essentially Boolean type
  - Fixed by having a boolean expression in the condition

+++ pm/drivers/clock/clock.c
@@ -3,7 +3,7 @@
  *
  * Cortex-M3 (CM3) firmware for power management
  *
- * Copyright (C) 2020, Texas Instruments Incorporated
+ * Copyright (C) 2020-2023, Texas Instruments Incorporated
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -42,6 +42,7 @@
 #include <types/errno.h>
 #include <osal/osal_core.h>
 #include <tisci/tisci_protocol.h>
+#include <pm.h>
 
 #include <tisci/pm/tisci_pm_clock.h>
 #include <limits.h>
@@ -369,18 +370,18 @@ s32 set_freq_handler(u32 *msg_recv)
 
 	ret = device_prepare_exclusive(req->hdr.host, id, NULL, &dev);
 	if (ret == SUCCESS) {
-		if ((min_freq_hz > ULONG_MAX) || (min_freq_hz > target_freq_hz) ||
+		if ((min_freq_hz > (u64) ULONG_MAX) || (min_freq_hz > target_freq_hz) ||
 		    (target_freq_hz > max_freq_hz)) {
 			ret = -EINVAL;
 		}
 	}
 
 	if (ret == SUCCESS) {
-		if (max_freq_hz > ULONG_MAX) {
+		if (max_freq_hz > (u64) ULONG_MAX) {
 			max_freq_hz = ULONG_MAX;
 		}
 
-		if (target_freq_hz > ULONG_MAX) {
+		if (target_freq_hz > (u64) ULONG_MAX) {
 			target_freq_hz = ULONG_MAX;
 		}
 
@@ -427,18 +428,18 @@ s32 query_freq_handler(u32 *msg_recv)
 
 	ret = device_prepare_exclusive(req->hdr.host, id, NULL, &dev);
 	if (ret == SUCCESS) {
-		if ((min_freq_hz > ULONG_MAX) || (min_freq_hz > target_freq_hz) ||
+		if ((min_freq_hz > (u64) ULONG_MAX) || (min_freq_hz > target_freq_hz) ||
 		    (target_freq_hz > max_freq_hz)) {
 			ret = -EINVAL;
 		}
 	}
 
 	if (ret == SUCCESS) {
-		if (max_freq_hz > ULONG_MAX) {
+		if (max_freq_hz > (u64) ULONG_MAX) {
 			max_freq_hz = ULONG_MAX;
 		}
 
-		if (target_freq_hz > ULONG_MAX) {
+		if (target_freq_hz > (u64) ULONG_MAX) {
 			target_freq_hz = ULONG_MAX;
 		}
